+++ default/role.harvester.js
@@ -14,9 +14,10 @@ module.exports = {
       If charged, transfer Energy to Structures
       */
     if (creep.isCharged()) {
-      
-      // Include Towers
-      creep.rechargeStructures(true);
+
+      // 25% chances the Hauler will also take care of Towers
+      if (creep.memory.includeTowers === undefined) creep.memory.includeTowers = _.sample([true, false, false, false]);
+      creep.rechargeStructures(creep.memory.includeTowers);
 
     /**
       Else, go recharging