+++ default/role.hero.js
@@ -72,7 +72,7 @@ module.exports = {
         Else, creep should be in homeroom
         */
       } else {
-        
+
         // Do not recharge Towers
         creep.rechargeStructures(false);
       }