+++ default/role.miner.js
@@ -4,12 +4,24 @@ module.exports = {
 
     // creep.say("M");
 
+    // TO DO: Store container and source object
+    // so they don't have to be searched all over again and again
+
     /**
       Miners get nearby an extraction point and focus on mining,
       Resources being mined get dropped into a nearby container.
       */
-    if (creep.isLocked() && creep.isCharged())  creep.transfer(creep.memory.container, RESOURCE_ENERGY);
-    if (creep.isLocked() && !creep.isCharged()) creep.harvest(creep.memory.source);
+    if (creep.isLocked() && creep.isCharged()) {
+      creep.memory.container = creep.pos.findInRange(FIND_STRUCTURES, 1, {
+        filter: s => s.structureType == STRUCTURE_CONTAINER
+      })[0];
+      creep.transfer(creep.memory.container, RESOURCE_ENERGY)
+    }
+    if (creep.isLocked() && !creep.isCharged()) {
+      // console.log(creep.pos.findInRange(FIND_SOURCES, 1), creep.pos.findClosestByRange(FIND_SOURCES));
+      creep.memory.source = creep.pos.findClosestByRange(FIND_SOURCES);
+      creep.harvest(creep.memory.source);
+    }
 
     /**
       Miners need to get close to an extraction point in order to lock on it